refactor: Move http information annotations to data.http
