+++ src/winter/data/client/FormData.java
@@ -4,7 +4,7 @@ import java.lang.reflect.Parameter;
 import java.util.HashMap;
 import java.util.Map;
 
-import winter.data.annotation.methodlevel.RequestParam;
+import winter.data.annotation.http.RequestParam;
 import winter.exception.AnnotationNotFoundException;
 
 public class FormData {
