+++ src/winter/util/AnnotationScanner.java
@@ -12,7 +12,7 @@ import jakarta.servlet.ServletContext;
 import winter.data.Mapping;
 import winter.data.MappingMethod;
 import winter.data.annotation.classlevel.Controller;
-import winter.data.annotation.methodlevel.UrlMapping;
+import winter.data.annotation.http.UrlMapping;
 import winter.exception.DuplicateMappingException;
 import winter.exception.InvalidPackageNameException;
 import winter.exception.PackageProviderNotFoundException;
